+++ docker-compose.yml
@@ -108,6 +108,19 @@ services:
       # - PYTHONUNBUFFERED=1
     ports:
       - 5009:5009
+
+  core_notification_complex:
+    container_name: core_notification_complex
+    build:
+      context: ./services/backend/core
+      dockerfile: Dockerfile-notification-complex-dev
+    networks:
+      - tracktrace_test_network
+    environment:
+      - SQLALCHEMY_DATABASE_URI=${DB_URI}
+      # - PYTHONUNBUFFERED=1
+    ports:
+      - 5015:5015
   
   core_view_all:
     container_name: core_view_all
