+++ services/backend/core/complex_scraper.py
@@ -39,138 +39,138 @@ db = SQLAlchemy(app)
 def health_check():
     return("complex_scraper")
 
-@app.route("/addsubscription", methods=['POST'])
-def addsubscription():
-     if request.is_json:
-        try:
-            data = request.get_json()
-            user_id = data["userid"]
-            containerid=data["containerid"]
-            status=data["status"]
+# @app.route("/addsubscription", methods=['POST'])
+# def addsubscription():
+#      if request.is_json:
+#         try:
+#             data = request.get_json()
+#             user_id = data["userid"]
+#             containerid=data["containerid"]
+#             status=data["status"]
 
-            jsondata={
-                "userid":user_id,
-                "containerid":containerid,
-                "status":status
-            }
-            response = invoke_http2("core_subscription", "subscription/add",prod, method='POST', json=jsondata)
-            print(response)
-        except Exception as e:
-            return jsonify(
-            {
-                "code": 500,
-                "message": e
-            }
-        ), 500
+#             jsondata={
+#                 "userid":user_id,
+#                 "containerid":containerid,
+#                 "status":status
+#             }
+#             response = invoke_http2("core_subscription", "subscription/add",prod, method='POST', json=jsondata)
+#             print(response)
+#         except Exception as e:
+#             return jsonify(
+#             {
+#                 "code": 500,
+#                 "message": e
+#             }
+#         ), 500
     
-        return response
+#         return response
 
-@app.route("/deletesubscription", methods=['POST'])
-def deletesubscription():
-     if request.is_json:
-        try:
-            data = request.get_json()
-            containerid=data["containerid"]
-            jsondata={
+# @app.route("/deletesubscription", methods=['POST'])
+# def deletesubscription():
+#      if request.is_json:
+#         try:
+#             data = request.get_json()
+#             containerid=data["containerid"]
+#             jsondata={
      
-                "containerid":containerid
+#                 "containerid":containerid
              
-            }
-            response = invoke_http2("core_subscription", "subscription/delete",prod, method='POST', json=jsondata)
-            print(response)
-        except Exception as e:
-            return jsonify(
-            {
-                "code": 500,
-                "message": e
-            }
-        ), 500
+#             }
+#             response = invoke_http2("core_subscription", "subscription/delete",prod, method='POST', json=jsondata)
+#             print(response)
+#         except Exception as e:
+#             return jsonify(
+#             {
+#                 "code": 500,
+#                 "message": e
+#             }
+#         ), 500
     
-        return response
-@app.route("/sendsms", methods=['POST'])
-def sendsms():
-    response = invoke_http2("core_subscription", "subscription/getsubscriptions",prod, method='POST')
-    for i in range(len(response)):
-        container=response[i]["container_id"]
-        userid=response[i]["wguser_id"]
-        subscription_status=response[i]["status"]
-        shipment_type="ctr"
-        directions="import"
-        data={
-            "identifier":container,
-            "identifier_type":shipment_type,
-            "direction":directions
+#         return response
+# @app.route("/sendsms", methods=['POST'])
+# def sendsms():
+#     response = invoke_http2("core_subscription", "subscription/getsubscriptions",prod, method='POST')
+#     for i in range(len(response)):
+#         container=response[i]["container_id"]
+#         userid=response[i]["wguser_id"]
+#         subscription_status=response[i]["status"]
+#         shipment_type="ctr"
+#         directions="import"
+#         data={
+#             "identifier":container,
+#             "identifier_type":shipment_type,
+#             "direction":directions
 
-        }
-        identifier = container
-        identifier_type = shipment_type
-        direction = directions
+#         }
+#         identifier = container
+#         identifier_type = shipment_type
+#         direction = directions
 
-            # print("\nReceived details in JSON:", data)
+#             # print("\nReceived details in JSON:", data)
 
-        if direction == "export":
-            origin = "SINGAPORE"
+#         if direction == "export":
+#             origin = "SINGAPORE"
 
-        # Retrieve Master BL from House BL
-        if identifier_type == "bl":
-            if direction == "import":
-                master_bl, origin, import_ref_n = get_import_master_bl(identifier)
-            elif direction == "export":
-                master_bl, export_ref_n = get_export_master_bl(identifier)            
+#         # Retrieve Master BL from House BL
+#         if identifier_type == "bl":
+#             if direction == "import":
+#                 master_bl, origin, import_ref_n = get_import_master_bl(identifier)
+#             elif direction == "export":
+#                 master_bl, export_ref_n = get_export_master_bl(identifier)            
 
-            data = {
-                        "identifier": master_bl,
-                        "identifier_type": "bl"
-                    }
+#             data = {
+#                         "identifier": master_bl,
+#                         "identifier_type": "bl"
+#                     }
         
-        if identifier_type == "ctr":
-            if direction == "import":
-                master_bl, origin, import_ref_n = get_import_master_bl_ctr(identifier)
-            elif direction == "export":
-                master_bl, export_ref_n = get_export_master_bl_ctr(identifier)   
+#         if identifier_type == "ctr":
+#             if direction == "import":
+#                 master_bl, origin, import_ref_n = get_import_master_bl_ctr(identifier)
+#             elif direction == "export":
+#                 master_bl, export_ref_n = get_export_master_bl_ctr(identifier)   
         
-        # Retrieve shipping line's prefix
-        prefix, vendor_name = get_prefix(master_bl, direction)
+#         # Retrieve shipping line's prefix
+#         prefix, vendor_name = get_prefix(master_bl, direction)
 
-        # Invoke scraper microservice
-        # shipment_info = invoke_http(scraper_url + prefix, method='POST', json=data)
-        shipment_info = invoke_http2("scraper_"+ prefix, prefix, prod, method="POST", json=data)
-        status=""
-        if shipment_info:
+#         # Invoke scraper microservice
+#         # shipment_info = invoke_http(scraper_url + prefix, method='POST', json=data)
+#         shipment_info = invoke_http2("scraper_"+ prefix, prefix, prod, method="POST", json=data)
+#         status=""
+#         if shipment_info:
         
-            status = shipment_info["data"]["status"]
+#             status = shipment_info["data"]["status"]
 
-            # Update DB with latest shipment information
+#             # Update DB with latest shipment information
             
 
-        response2=status
+#         response2=status
         
-        if(subscription_status!=response2):
-            updatedata={
-                "containerid":identifier,
-                "status":response2
-            }
+#         if(subscription_status!=response2):
+#             updatedata={
+#                 "containerid":identifier,
+#                 "status":response2
+#             }
             
-            useriddata={
-            "wguserid":userid
-            }
-            phone_number=invoke_http2("core_user", "user/getnumber",prod, method='POST', json=useriddata)
+#             useriddata={
+#             "wguserid":userid
+#             }
+#             phone_number=invoke_http2("core_user", "user/getnumber",prod, method='POST', json=useriddata)
             
-            account_sid = "AC7a7b489784baef97d21697b086b468ec"
-            auth_token = "eed37e9b4744d26ed34a67d4d3efe232"
-            client = Client(account_sid, auth_token)
-            message = client.messages.create(
-                body="There has been a status update in regards to your Container Number: "+identifier+" Status:"+response2,
-                from_="+15674004435",
-                to = "+65"+str(phone_number)
-                )
-            update = invoke_http2("core_subscription", "subscription/update",prod, method='POST',json=updatedata)
+#             account_sid = "AC7a7b489784baef97d21697b086b468ec"
+#             auth_token = "eed37e9b4744d26ed34a67d4d3efe232"
+#             client = Client(account_sid, auth_token)
+#             message = client.messages.create(
+#                 body="There has been a status update in regards to your Container Number: "+identifier+" Status:"+response2,
+#                 from_="+15674004435",
+#                 to = "+65"+str(phone_number)
+#                 )
+#             update = invoke_http2("core_subscription", "subscription/update",prod, method='POST',json=updatedata)
         
 
-    return jsonify({
-                "code": 200,
-                "message": "success"
-            }),200
+#     return jsonify({
+#                 "code": 200,
+#                 "message": "success"
+#             }),200
 
 @app.route('/scrape', methods=['POST'])
 def scrape():
