+++ services/frontend/app/src/api/config.js
@@ -5,6 +5,7 @@ const { REACT_APP_API_ENDPOINT } = process.env;
 export const USER_ENDPOINT = `http://${REACT_APP_API_ENDPOINT}:5002/user`;
 export const COMPLEX_SCRAPER_ENDPOINT = `http://${REACT_APP_API_ENDPOINT}:5009`;
 export const VIEW_ALL_SHIPMENTS_ENDPOINT = `http://${REACT_APP_API_ENDPOINT}:5010`;
+export const NOTIFICATION_COMPLEX_ENDPOINT = `http://${REACT_APP_API_ENDPOINT}:5015`;
 
 // to extract csrf_access_token, which should then be put into request header ("X-CSRF-TOKEN")
 export function getCookie(name) {
